TERMINAR EL PINCHE SPLIT -________-(palabras)

+++ ft_strlcpy.c
@@ -6,10 +6,11 @@
 /*   By: mcuesta- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 16:05:20 by mcuesta-          #+#    #+#             */
-/*   Updated: 2024/09/18 15:12:58 by mcuesta-         ###   ########.fr       */
+/*   Updated: 2024/10/02 11:59:27 by mcuesta-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <stddef.h>
+
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	int	i;
@@ -17,7 +18,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 
 	i = 0;
 	j = 0;
-	while (src[i] != '\n' && i < size)
+	while (src[i] != '\n' && i < (int)size)
 	{
 		dst[j] = src[i];
 		i++;
